requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
numpy==2.0.0
opencv-python-headless==4.10.0.84
onnxruntime==1.19.0